mediai-diagnostics/backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
pillow
pytesseract
